style: rename tip1028_blocked_transfers module to tip1028_guard
Amp-Thread-ID: https://ampcode.com/threads/T-019e269e-ef4c-77f9-a1b7-674f9961dd30
