+++ crates/contracts/src/precompiles/tip1028_guard.rs
