+++ xtask/src/genesis_args.rs
@@ -59,7 +59,7 @@ use tempo_precompiles::{
     tip20::{ISSUER_ROLE, ITIP20, TIP20Token},
     tip20_factory::TIP20Factory,
     tip403_registry::TIP403Registry,
-    tip1028_blocked_transfers::TIP1028Guard,
+    tip1028_guard::TIP1028Guard,
     validator_config_v2::ValidatorConfigV2,
 };
 
@@ -421,7 +421,7 @@ impl GenesisArgs {
 
         if self.t6_time == 0 {
             println!("Initializing TIP1028 blocked transfers (T6 active at genesis)");
-            initialize_tip1028_blocked_transfers(&mut evm)?;
+            initialize_tip1028_guard(&mut evm)?;
         }
 
         if !self.no_pairwise_liquidity {
@@ -936,7 +936,7 @@ fn initialize_signature_verifier(evm: &mut TempoEvm<CacheDB<EmptyDB>>) -> eyre::
     Ok(())
 }
 
-fn initialize_tip1028_blocked_transfers(evm: &mut TempoEvm<CacheDB<EmptyDB>>) -> eyre::Result<()> {
+fn initialize_tip1028_guard(evm: &mut TempoEvm<CacheDB<EmptyDB>>) -> eyre::Result<()> {
     let ctx = evm.ctx_mut();
     StorageCtx::enter_evm(
         &mut ctx.journaled_state,
